refactor: moved command execution into IProgramOperator
