+++ Piet.Interpreter/IProgramOperator.cs
@@ -1,0 +1,8 @@
+﻿using Piet.Command;
+
+namespace Piet.Interpreter;
+
+public interface IProgramOperator
+{
+    void ExecuteCommand(ColorCommand colorCommand, int codelBlockSize);
+}
