+++ Piet.Interpreter/PietInterpreter.cs
@@ -1,8 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
 using Piet.Command;
 using Piet.Grid;
-using Piet.Interpreter.Events;
-using Piet.Interpreter.Exceptions;
 
 namespace Piet.Interpreter;
 
@@ -10,11 +8,9 @@ public sealed class PietInterpreter
 {
     private readonly ICodelChooser _codelChooser;
     private readonly ICodelBlockSearcher _codelBlockSearcher;
-    private readonly IOutputEventService _outputEventService;
-    private readonly IInputService _inputService;
+    private readonly IProgramOperator _programOperator;
     private readonly ILogger<PietInterpreter> _logger;
     
-    private readonly Stack<int> _programStack;
     private Codel _currentCodel;
     private bool _executionFinished = false;
     private bool _executionError = false;
@@ -27,29 +23,15 @@ public sealed class PietInterpreter
         ICodelGrid codelGrid,
         ICodelChooser codelChooser,
         ICodelBlockSearcher codelBlockSearcher,
-        IOutputEventService outputEventService,
-        IInputService inputService
+        IProgramOperator programOperator
     )
     {
         _logger = logger;
         _codelChooser = codelChooser;
         _codelBlockSearcher = codelBlockSearcher;
-        _outputEventService = outputEventService;
-        _inputService = inputService;
-        _programStack = new();
+        _programOperator = programOperator;
         _currentCodel = codelGrid.GetCodel(0, 0);
     }
-
-    public void TEST_TriggerOutputOperation(int value)
-    {
-        _outputEventService.DispatchOutputIntegerEvent(value);
-    }
-
-    public void TEST_TriggerOutputOperation(char value)
-    {
-        _outputEventService.DispatchOutputCharacterEvent(value);
-    }
-
 
     public void Run()
     {
@@ -68,7 +50,7 @@ public sealed class PietInterpreter
         }
         else
         {
-            _logger.LogCritical("An error occured - Piet interpreter stopped");
+            _logger.LogCritical("An error occurred - Piet interpreter stopped");
         }
     }
 
@@ -76,7 +58,7 @@ public sealed class PietInterpreter
     {
         _logger.LogDebug("Get codel block for current codel: {_currentCodel}", _currentCodel);
         var codelBock = _codelBlockSearcher.GetCodelBock(_currentCodel).ToList();
-        _logger.LogDebug($"Retrived codel block of size {codelBock.Count()}");
+        _logger.LogDebug($"Retrieved codel block of size {codelBock.Count()}");
         
         _logger.LogDebug("Determine next codel");
         var nextCodelResult = _codelChooser.GetNextCodel(codelBock);
@@ -96,7 +78,7 @@ public sealed class PietInterpreter
         if (nextCodelResult.TraversedWhiteCodels is false)
         {
             _logger.LogDebug($"Execute command: {colorCommand}");
-            ExecuteCommand(colorCommand, codelBock.Count());
+            _programOperator.ExecuteCommand(colorCommand, codelBock.Count());
             _logger.LogDebug($"Executed command: {colorCommand}");
         }
         
@@ -105,254 +87,7 @@ public sealed class PietInterpreter
     }
 
     private void UpdateCurrentCodel(Codel codel) => _currentCodel = codel;
-
-    private void ExecuteCommand(ColorCommand command, int codelBlockSize)
-    {
-        int operand;
-        int operandA;
-        int operandB;
-
-        //var test = command.Command switch
-        //{
-        //    Command.Command.None => NoneOperation(),
-        //    Command.Command.Add => Add(),
-        //    _ => throw new Exception()
-        //};
-        // TODO: create service CommandExecutionService which handel all commands
-
-
-        switch (command.Command)
-        {
-            case Command.Command.None:
-                _logger.LogDebug($"Executing command {Command.Command.None}: program state does not change");
-                break;
-            case Command.Command.Push:
-                _programStack.Push(codelBlockSize);
-                break;
-            case Command.Command.Pop:
-                _programStack.Pop();
-                break;
-            case Command.Command.Add:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operandB = _programStack.Pop();
-                operandA = _programStack.Pop();
-                _programStack.Push(operandA + operandB);
-                break;
-            case Command.Command.Subtract:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operandB = _programStack.Pop();
-                operandA = _programStack.Pop();
-                _programStack.Push(operandA - operandB);
-                break;
-            case Command.Command.Multiply:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operandB = _programStack.Pop();
-                operandA = _programStack.Pop();
-                _programStack.Push(operandA * operandB);
-                break;
-            case Command.Command.Divide:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operandB = _programStack.Pop();
-                operandA = _programStack.Pop();
-
-                if (operandB == 0)
-                {
-                    throw new PietInterpreterDividedByZeroException("Division by zero is undefined.");
-                }
-
-                _programStack.Push(operandA / operandB);
-                break;
-            case Command.Command.Modulo:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operandB = _programStack.Pop();
-                operandA = _programStack.Pop();
-
-                if (operandB == 0)
-                {
-                    throw new PietInterpreterDividedByZeroException("Modulo division for zero is undefined.");
-                }
-
-                var result = operandA % operandB;
-                if (result < 0 && operandB > 0)
-                {
-                    result = Math.Abs(result);
-                }
-                _programStack.Push(result);
-                break;
-            case Command.Command.Not:
-                if (_programStack.Count < 1)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operand = _programStack.Pop();
-                if (operand == 0)
-                {
-                    _programStack.Push(1);
-                }
-                else
-                {
-                    _programStack.Push(0);
-                }
-
-                break;
-                
-            case Command.Command.Greater:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operandB = _programStack.Pop();
-                operandA = _programStack.Pop();
-
-                _programStack.Push(operandA > operandB ? 1 : 0);
-                break;
-
-            case Command.Command.Pointer:
-                if (_programStack.Count < 1)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operand = _programStack.Pop();
-
-                if (operand > 0)
-                {
-                    for (int i = 0; i < operand; i++)
-                    {
-                        RotateDirectionPointerClockwise();
-                    }
-                }
-
-                if (operand < 0)
-                {
-                    for (int i = 0; i < Math.Abs(operand); i++)
-                    {
-                        RotateDirectionPointerCounterClockwise();
-                    }
-                }
-                break;
-
-            case Command.Command.Switch:
-                if (_programStack.Count < 1)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operand = _programStack.Pop();
-
-                for (int i = 0; i <= Math.Abs(operand); i++)
-                {
-                    ToggleCodelChooser();
-                }
-
-                break;
-
-            case Command.Command.Duplicate:
-                if (_programStack.Count < 1)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operand = _programStack.Peek();
-                _programStack.Push(operand);
-                break;
-            
-            case Command.Command.Roll:
-                if (_programStack.Count < 2)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                var numberOfRolls = _programStack.Pop();
-                var depthOfRollOperation = _programStack.Pop();
-
-                // convert stack to array to perform roll operation
-                var stackAsArray = _programStack.ToArray();
-                Array.Reverse(stackAsArray);
-
-                if (depthOfRollOperation < stackAsArray.Length)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"Error in 'roll operation': There are {_programStack.Count} elements on the stack" +
-                        $"but a roll depth of {depthOfRollOperation} was requested.");
-                }
-
-                // roll
-                int rollInsertIndex = stackAsArray.Length - depthOfRollOperation;
-                for (int i = 0; i < numberOfRolls; i++)
-                {
-                    int programStackTopElement = stackAsArray[^1];
-                    Array.Copy(stackAsArray, rollInsertIndex, stackAsArray, rollInsertIndex + 1, stackAsArray.Length - rollInsertIndex - 1);
-                    stackAsArray.SetValue(programStackTopElement, rollInsertIndex);
-                }
-                
-                // back to stack
-                _programStack.Clear();
-                foreach (var number in stackAsArray)
-                {
-                    _programStack.Push(number);
-                }
-                
-                break;
-
-            case Command.Command.InputNumber:
-                int inputNumber = _inputService.GetIntegerInput().Result;
-                _programStack.Push(inputNumber);
-                break;
-            case Command.Command.InputCharacter:
-                int inputCharacter = _inputService.GetCharacterInput().Result;
-                _programStack.Push(inputCharacter);
-                break;
-            case Command.Command.OutputNumber:
-                if (_programStack.Count < 1)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operand = _programStack.Pop();
-                _logger.LogDebug($"Numeric output value {operand}");
-                _outputEventService.DispatchOutputIntegerEvent(operand);
-
-                break;
-            case Command.Command.OutputCharacter:
-                if (_programStack.Count < 1)
-                {
-                    throw new InsufficientNumberOfElementsOnProgramStackException($"There are {_programStack.Count} elements on the stack");
-                }
-
-                operand = _programStack.Pop();
-                _logger.LogDebug($"Character output value{Convert.ToChar(operand)}");
-                _outputEventService.DispatchOutputCharacterEvent((char)operand);
-
-                break;
-            default:
-                throw new ArgumentOutOfRangeException(
-                    $"The command ${command.Command} is not valid in this context");
-        }
-    }
-
-
+    
     internal static void ToggleCodelChooser()
     {
         CodelChooserState = CodelChooserState switch
